Subscription Id - long to int
Subscription id changed from long to int as per new api.

CRs-Fixed: 821673

+++ src/com/android/server/telecom/TelecomService.java
@@ -702,10 +702,10 @@ public class TelecomService extends Service {
         }
 
         @Override
-        public long getActiveSubscription() { return 0; }
+        public int getActiveSubscription() { return 0; }
 
         @Override
-        public void switchToOtherActiveSub(long subId) { }
+        public void switchToOtherActiveSub(int subId) { }
 
         /**
          * Dumps the current state of the TelecomService.  Used when generating problem reports.
